# docs/marketplace/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "KWIP Data Marketplace API",
    "version": "2.0.0",
    "description": "Public catalog, metered data delivery, scoped service keys, custom user-approved device linking, bounded Pro exports and operator push ingestion. Data is not seeded. This is HTTP API documentation, not an MCP server or OAuth specification. See /agents.txt. All JSON business responses use Cache-Control: no-store. Common peer ceiling 300 requests/minute; authenticated workspace default 120/minute, configurable. Honor Retry-After; metadata does not cost credits but is rate limited."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the served specification; production https://kwip.tech."
    }
  ],
  "paths": {
    "/api/v2/datasets": {
      "get": {
        "operationId": "listDatasets",
        "summary": "Explore public dataset metadata",
        "description": "Free metadata. Only published datasets with active approved sources appear. At most 50 per page, slug order. next_after is an opaque continuation value to pass as after. No row preview is returned.",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. Common statuses: 400 validation; 401 missing/invalid credential; 403 authority; 405 method; 413 size; 415 JSON media type; 429 throttled (Retry-After); 503 unavailable. Error codes are extensible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ]
      }
    },
    "/api/v2/datasets/{slug}": {
      "get": {
        "operationId": "getDataset",
        "summary": "Read schema, source attribution and per-record credits",
        "description": "",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetail"
                }
              }
            }
          },
          "404": {
            "description": "Dataset unavailable or unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. Common statuses: 400 validation; 401 missing/invalid credential; 403 authority; 405 method; 413 size; 415 JSON media type; 429 throttled (Retry-After); 503 unavailable. Error codes are extensible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[-a-zA-Z0-9_]{1,50}$"
            }
          }
        ]
      }
    },
    "/api/v2/plans": {
      "get": {
        "operationId": "getPlans",
        "summary": "Inspect configured allowances and purchase availability",
        "description": "Prices/allowances are server configuration. purchases_available is false until billing is enabled and approved published records exist; it is not proof of provider readiness.",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plans"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. Common statuses: 400 validation; 401 missing/invalid credential; 403 authority; 405 method; 413 size; 415 JSON media type; 429 throttled (Retry-After); 503 unavailable. Error codes are extensible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/billing/checkout": {
      "post": {
        "operationId": "startCheckout",
        "summary": "Open owner-approved Stripe Checkout",
        "description": "No request body or customer-selected price/customer IDs. Browser session only. Return URL is informational, not a paid entitlement. Owner completes Stripe action. Service keys are forbidden.",
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "additionalProperties": true,
                  "required": [
                    "url"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "catalog_empty/subscription_exists/billing_conflict (Checkout) or no_billing_account (Portal).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Billing disabled, misconfigured or provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. Common statuses: 400 validation; 401 missing/invalid credential; 403 authority; 405 method; 413 size; 415 JSON media type; 429 throttled (Retry-After); 503 unavailable. Error codes are extensible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/operator/datasets": {
      "post": {
        "operationId": "registerDataset",
        "summary": "Operator validates or registers dataset/schema",
        "description": "Operator browser session required. Defaults to dry_run=true. No dynamic SQL table. Register draft, separately register/approve sources, then publish. Schema versions immutable/additive. HTTP dataset object excludes sources; CLI combined manifests have a different wrapper. Slug identifies upsert; no Idempotency-Key used. Maximum HTTP body 1 MiB.",
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "pattern": "^[-a-zA-Z0-9_]{1,50}$"
                    },
                    "status": {
                      "type": "string"
                    },
                    "dry_run": {
                      "type": "boolean"
                    }
                  },
                  "additionalProperties": true,
                  "required": [
                    "slug",
                    "status",
                    "dry_run"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "JSON error. Common statuses: 400 validation; 401 missing/invalid credential; 403 authority; 405 method; 413 size; 415 JSON media type; 429 throttled (Retry-After); 503 unavailable. Error codes are extensible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataset": {
                    "$ref": "#/components/schemas/DatasetManifest"
                  },
                  "dry_run": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "additionalProperties": true,
                "required": [
                  "dataset"
                ]
              }
            }
          }
        },
        "x-required-scopes": [
          "ingest:write"
        ],
        "x-operator-required": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ServiceKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "kwip_ service key",
        "description": "Owner-created expiring/revocable service key; scopes datasets:read, exports:read, and operator-only ingest:write. Never sends key in URL."
      },
      "BrowserSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Clerk RS256 session JWT",
        "description": "Short-lived active Clerk user session token from the configured first-party browser sign-in. Management actions require this scheme, not service keys. No cookie-only API authentication."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retry_after": {
                "type": "integer",
                "minimum": 1
              },
              "required_credits": {
                "type": "integer",
                "minimum": 0
              }
            },
            "additionalProperties": true,
            "required": [
              "code"
            ]
          }
        },
        "additionalProperties": true,
        "required": [
          "error"
        ]
      },
...
      "FieldDefinition": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "string",
              "integer",
              "number",
              "boolean",
              "datetime",
              "date",
              "enum"
            ]
          },
          "required": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "maxLength": 10000
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 100000
            },
            "minItems": 1,
            "maxItems": 100
          }
        },
        "additionalProperties": false,
        "required": [
          "type"
        ]
      },
      "FieldMap": {
        "type": "object",
        "minProperties": 1,
        "maxProperties": 100,
        "propertyNames": {
          "pattern": "^(?!.*__)[a-z][a-z0-9_]{0,62}$"
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/FieldDefinition"
        }
      },
...
      "PublicSource": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "license_url": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "required": [
          "slug",
          "name",
          "url",
          "attribution",
          "license_url"
        ]
      },
      "Dataset": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[-a-zA-Z0-9_]{1,50}$"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "credits_per_record": {
            "type": "integer",
            "minimum": 1
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "required": [
          "slug",
          "title",
          "description",
          "category",
          "credits_per_record",
          "updated_at",
          "url",
          "schema_version"
        ]
      },
      "DatasetDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Dataset"
          },
          {
            "type": "object",
            "properties": {
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/FieldDefinition"
                }
              },
              "sources": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PublicSource"
                },
                "maxItems": 100
              },
              "limits": {
                "type": "object",
                "properties": {
                  "page_size": {
                    "type": "integer"
                  },
                  "filters": {
                    "type": "integer"
                  },
                  "cursor_lifetime_seconds": {
                    "type": "integer"
                  }
                },
                "additionalProperties": true,
                "required": [
                  "page_size",
                  "filters",
                  "cursor_lifetime_seconds"
                ]
              }
            },
            "additionalProperties": true,
            "required": [
              "fields",
              "sources",
              "limits"
            ]
          }
        ]
      },
      "Catalog": {
        "type": "object",
        "properties": {
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dataset"
            },
            "maxItems": 50
          },
          "next_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "required": [
          "datasets",
          "next_after"
        ]
      },
      "Plans": {
        "type": "object",
        "properties": {
          "free_credits": {
            "type": "integer",
            "minimum": 0
          },
          "pro_credits": {
            "type": "integer",
            "minimum": 0
          },
          "pro_monthly_cents": {
            "type": "integer",
            "minimum": 0
          },
          "overage_cents_per_10000": {
            "type": "integer",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd"
            ]
          },
          "purchases_available": {
            "type": "boolean"
          },
          "catalog_has_data": {
            "type": "boolean"
          }
        },
        "additionalProperties": true,
        "required": [
          "free_credits",
          "pro_credits",
          "pro_monthly_cents",
          "overage_cents_per_10000",
          "currency",
          "purchases_available",
          "catalog_has_data"
        ]
      },
...
      "DatasetManifest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[-a-zA-Z0-9_]{1,50}$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "withdrawn"
            ],
            "default": "draft"
          },
          "credits_per_record": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "default": 1
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "fields": {
            "$ref": "#/components/schemas/FieldMap"
          }
        },
        "additionalProperties": false,
        "required": [
          "slug",
          "title",
          "schema_version",
          "fields"
        ]
      },
...
    }
  }
}